requirements.txt
trulens-eval~=0.27.2
scikit-learn~=1.4.2
sentence-transformers~=2.6.1
langchain-openai~=0.1.3
unstructured~=0.13.2
unstructured-client~=0.18.0
llama-index~=0.10.29
llama-index-postprocessor-cohere-rerank~=0.1.4
llama-index-embeddings-together~=0.1.3
llama-index-llms-together~=0.1.3
llama-index-vector-stores-weaviate~=0.1.4